@@ -11,32 +11,38 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^8.0.1",
-    "@angular/cdk": "^8.0.1",
-    "@angular/common": "~8.0.0",
-    "@angular/compiler": "~8.0.0",
-    "@angular/core": "~8.0.0",
+    "@angular/animations": "^9.1.0",
+    "@angular/cdk": "^9.2.0",
+    "@angular/common": "~9.1.0",
+    "@angular/compiler": "~9.1.0",
+    "@angular/core": "~9.1.0",
     "@angular/fire": "^6.0.0",
-    "@angular/forms": "~8.0.0",
-    "@angular/material": "^8.0.1",
-    "@angular/platform-browser": "~8.0.0",
-    "@angular/platform-browser-dynamic": "~8.0.0",
-    "@angular/router": "~8.0.0",
+    "@angular/forms": "~9.1.0",
+    "@angular/material": "^9.2.0",
+    "@angular/platform-browser": "~9.1.0",
+    "@angular/platform-browser-dynamic": "~9.1.0",
+    "@angular/router": "~9.1.0",
+    "angular-bootstrap-datetimepicker": "^4.0.2",
+    "angular-mat-datepicker": "0.0.2",
+    "angular-material": "^1.1.21",
+    "angular-material-datepicker": "^1.0.2",
+    "bootstrap": "^4.4.1",
     "firebase": "^7.13.2",
-    "hammerjs": "^2.0.8",
-    "rxjs": "~6.4.0",
-    "tslib": "^1.9.0",
-    "zone.js": "~0.9.1"
+    "moment": "^2.27.0",
+    "open-iconic": "^1.1.1",
+    "rxjs": "~6.5.5",
+    "tslib": "^1.10.0",
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.800.0",
-    "@angular/cli": "~8.0.2",
-    "@angular/compiler-cli": "~8.0.0",
-    "@angular/language-service": "~8.0.0",
-    "@types/node": "~8.9.4",
+    "@angular-devkit/build-angular": "~0.901.0",
+    "@angular/cli": "~9.1.0",
+    "@angular/compiler-cli": "~9.1.0",
+    "@angular/language-service": "~9.1.0",
+    "@types/node": "^12.11.1",
     "@types/jasmine": "~3.3.8",
     "@types/jasminewd2": "~2.0.3",
-    "codelyzer": "^5.0.0",
+    "codelyzer": "^5.1.2",
     "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~4.1.0",
@@ -47,7 +53,7 @@
     "protractor": "~5.4.0",
     "ts-node": "~7.0.0",
     "tslint": "~5.15.0",
-    "typescript": "~3.4.3",
+    "typescript": "~3.8.3",
     "@angular-devkit/architect": "~0.900",
     "firebase-tools": "^7.12.1",
     "fuzzy": "^0.1.3",
