@@ -50,7 +50,7 @@ export class PolicyListComponent implements OnInit {
       this.policies = data.map(e => {
         return {
           id: e.payload.doc.id,
-          ...e.payload.doc.data()
+          ...(e.payload.doc.data() as object)
         } as Policy;
       })
     });
